make folder for helper files, and get battleships and tic-tac-toe to use it
