+++ community/games/README.md
@@ -9,9 +9,9 @@ This is exactly what we'd like to replicate in this folder. Here you'll find bas
 
 * [Battleships with partial NOT gates](battleships_with_partial_NOT_gates.ipynb)
 
-* [Quantum Counterfeit Coin Problem]()
+* [Quantum Counterfeit Coin Problem](quantum_counterfeit_coin_problem.ipynb)
 
-* [Quantum Tic-tac-toe](quantum_tic_tac_toe)
+* [Quantum Tic-tac-toe](quantum_tic_tac_toe.ipynb)
 
 
 ## Contributing
@@ -20,7 +20,7 @@ Each game should have a Jupyter notebook in which the game is played, and in whi
 
 If you are intimidated by the idea of explaining the quantum program, don't hesitate to contribute your game without it. Others will be happy to build on your work, and bring it in line with the required style.
 
-To make even a simple command line game, you'll most likely need many lines of classical programming. This will handle the subtelties of input and output, and so not be directly relevant to someone learning quantum programming. You can therefore hide this away in a `.py` file that can then be imported into your notebook. For example see `battleships_engine.py` and its use in [Battleships with partial NOT gates](battleships_with_partial_NOT_gates.ipynb).
+To make even a simple command line game, you'll most likely need many lines of classical programming. This will handle the subtelties of input and output, and so not be directly relevant to someone learning quantum programming. You can therefore hide this away in a _.py_ file that can then be imported into your notebook. These files can be placed in the _game_engines_ folder. For example see _battleships_engine.py_ and its use in [Battleships with partial NOT gates](battleships_with_partial_NOT_gates.ipynb).
 
 For more on how to contribute, see the main [README](../../README.md) for these tutorials.
